@@ -21,142 +21,34 @@
 		"firefox": "browser=firefox webpack",
 		"clean": "rm -rf ./dist && rm -rf ./build",
 		"test": "jest",
-		"build": "npm run clean &&  npm run chrome",
-		"rome": "rome format ."
-	},
-	"resolutions": {
-		"babel-jest": "^21.2.0",
-		"babel-plugin-jest-hoist": "^21.2.0",
-		"babel-preset-jest": "^21.2.0"
+		"build": "npm run clean &&  npm run chrome"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.11.6",
-		"@babel/plugin-proposal-optional-chaining": "^7.11.0",
-		"@babel/preset-env": "^7.11.5",
-		"@tailwindcss/aspect-ratio": "^0.4.0",
-		"@tailwindcss/forms": "^0.4.0",
-		"@tailwindcss/line-clamp": "^0.3.1",
-		"@tailwindcss/typography": "^0.5.2",
-		"autoprefixer": "^10.4.2",
-		"babel-eslint": "^10.1.0",
-		"css-loader": "^4.3.0",
-		"eslint": "^7.32.0",
-		"eslint-plugin-vue": "^7.20.0",
-		"eslint-plugin-vuejs-accessibility": "^0.5.1",
-		"expose-loader": "^3.0.0",
-		"file-loader": "^6.1.1",
-		"fs": "0.0.2",
-		"html-minimizer-webpack-plugin": "^3.3.0",
-		"jest": "^26.5.3",
-		"jest-chrome": "^0.7.1",
-		"konami-code": "^0.3.0",
-		"mark.js": "^8.11.1",
-		"medium-button": "^1.1.4",
-		"postcss": "^8.4.7",
-		"postcss-loader": "^4.1.0",
+		"@babel/core": "^7.22.9",
+		"@babel/preset-env": "^7.22.9",
+		"@babel/preset-react": "^7.22.5",
+		"@meforma/vue-toaster": "^1.3.0",
+		"archiver": "^5.3.1",
+		"autoprefixer": "^10.4.14",
+		"babel-loader": "^9.1.3",
+		"copy-webpack-plugin": "^11.0.0",
+		"css-loader": "^6.8.1",
+		"html2canvas": "^1.4.1",
+		"postcss": "^8.4.26",
+		"postcss-loader": "^7.3.3",
+		"style-loader": "^3.3.3",
+		"stylus-loader": "^7.1.3",
+		"tailwindcss": "^3.3.3",
+		"vue-loader": "^17.2.2",
 		"raw-loader": "^4.0.2",
-		"rome": "^0.8.0-next.ff4153b",
-		"style-loader": "^2.0.0",
-		"tailwindcss": "^3.0.23",
-		"tesseract.js": "^2.1.4",
-		"vue-loader": "^15.9.3",
-		"webpack": "^4.44.1",
-		"webpack-bundle-analyzer": "^4.5.0",
-		"webpack-cli": "^4.2.0",
-		"zip-webpack-plugin": "^3.0.0"
-	},
-	"eslintConfig": {
-		"root": true,
-		"extends": [
-			"plugin:vue/essential",
-			"plugin:prettier/recommended",
-			"eslint:recommended"
-		],
-		"env": {
-			"browser": true,
-			"jest": true,
-			"node": true
-		},
-		"globals": {
-			"chrome": true
-		},
-		"rules": {
-			"eol-last": 0,
-			"quotes": [
-				2,
-				"single"
-			]
-		}
+		"webpack": "^5.88.1",
+		"webpack-cli": "^5.1.4",
+		"zip-webpack-plugin": "^4.0.1"
 	},
 	"dependencies": {
-		"@microsoft/immersive-reader-sdk": "^1.2.0",
-		"@mozilla/readability": "^0.4.2",
-		"@postlight/mercury-parser": "^2.2.1",
-		"async": "^3.2.0",
-		"axios": "^0.20.0",
-		"axios-jsonp": "^1.0.4",
-		"babel-loader": "^8.2.2",
-		"cli-strip-html": "^1.0.2",
-		"config-chain": "^1.1.13",
-		"copy-webpack-plugin": "^6.2.1",
-		"country-code-emoji": "^2.3.0",
-		"daisyui": "^2.4.0",
-		"eslint-config-prettier": "^6.12.0",
-		"eslint-plugin-prettier": "^3.1.4",
-		"fetch-jsonp": "^1.2.1",
-		"get-json": "^1.0.1",
-		"google-docs-utils": "^2.3.0",
-		"google-translate": "^3.0.0",
-		"he": "^1.2.0",
-		"hotkeys-js": "^3.8.7",
-		"html-entities": "^2.3.3",
-		"html-to-markdown": "^1.0.0",
-		"jcrop": "^3.0.1",
-		"jcrop-0.9.12": "0.9.12",
-		"jest-puppeteer": "^6.1.0",
-		"jquery": "^3.6.0",
-		"js-confetti": "^0.9.0",
-		"jsdom": "^16.4.0",
-		"jsonp": "^0.2.1",
-		"konami": "^1.6.3",
-		"markdown": "^0.5.0",
-		"material-icons": "^0.3.1",
-		"mathsteps": "^0.2.0",
-		"medium-editor": "^5.23.3",
-		"moment-mini": "^2.24.0",
-		"moment-timezone": "^0.5.34",
-		"node-html-markdown": "^1.1.3",
-		"notyf": "^3.10.0",
-		"npm": "^6.14.8",
-		"npmlog": "^6.0.0",
-		"path": "^0.12.7",
-		"pdfjs-dist": "^2.5.207",
-		"pdfvuer": "^1.9.2",
-		"predictionary": "^1.4.4",
-		"prettier": "^2.1.2",
-		"pretty": "^2.0.0",
-		"remove-markdown": "^0.3.0",
-		"resize-observer-polyfill": "^1.5.1",
-		"resolve-url-loader": "^4.0.0",
-		"strip-markdown": "^5.0.0",
-		"tesseract.js-core": "^2.2.0",
-		"traverse": "^0.6.6",
-		"turndown": "^7.1.1",
-		"vue": "^2.6.12",
-		"vue-async-computed": "^3.9.0",
-		"vue-color": "^2.8.1",
-		"vue-konami-code": "^1.0.0",
-		"vue-pdf": "^4.3.0",
-		"vue-plugin-webextension-i18n": "^0.1.2",
-		"vue-select": "^3.18.3",
-		"vue-slider-component": "^3.2.6",
-		"vue-spinner": "^1.0.4",
-		"vue-swatches": "^2.1.0",
-		"vue-template-compiler": "^2.6.12",
-		"vue-tour": "^2.0.0",
-		"vue2-editor": "^2.10.2",
-		"vuelidate": "^0.7.6",
-		"web-highlighter": "^0.7.1",
-		"yargs": "^16.2.0"
+		"daisyui": "^3.1.6",
+		"toastify-js": "^1.12.0",
+		"vue": "^3.3.4",
+		"vue-loader": "^17.2.2"
 	}
 }
